Feature added to pass header object with key and value of header
Often times the data object keys and the header names are not same. So I've
added option to pass header in the form of object. The keys would be same as
that of the keys of the data to be shown and the value would be used as the header of the column for that key. Also the sequence of keys don't matter.
They will be ordered according the the header object passed.